Tests/LibWeb/Text/expected/top-layer-nested-fullscreen-members-keep-painting.txt

1: first
2: last
3: first
painted
